requirements.txt
numpy
pandas
pymc
arviz
pytensor
pyyaml
matplotlib
